requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
sgmllib3k==1.0.0
tkfontchooser==2.0.2
